requirements.txt
numpy==1.25.2
scipy==1.12.0
pysindy==1.7.5
pyyaml==6.0.1
matplotlib==3.7.1
pandas==2.1.4
ipykernel==6.29.4 # For .ipynb execution.
Jinja2 # For Python dataframe.style